refactor and add progress to every process
