+++ src/main/java/com/desktop/services/processors/interfaces/IDocumentProcess.java
@@ -6,7 +6,5 @@ import org.jsoup.nodes.Document;
 import java.util.concurrent.CompletableFuture;
 
 public interface IDocumentProcess {
-    CompletableFuture<Void> ProcessAsync(Document document);
-
     CompletableFuture<Void> ProcessAsync(Document document, DoubleProperty progress);
 }
